Add SQF for watched images and benchmark tests
