+++ central/reports/scheduler/v2/reportgenerator/singleton.go
@@ -4,6 +4,7 @@ import (
 	"github.com/graph-gophers/graphql-go"
 	blobDS "github.com/stackrox/rox/central/blob/datastore"
 	clusterDS "github.com/stackrox/rox/central/cluster/datastore"
+	imageCVEDS "github.com/stackrox/rox/central/cve/image/datastore"
 	deploymentDS "github.com/stackrox/rox/central/deployment/datastore"
 	"github.com/stackrox/rox/central/globaldb"
 	"github.com/stackrox/rox/central/graphql/resolvers"
@@ -35,6 +36,7 @@ func initialize() {
 		blobDS.Singleton(),
 		clusterDS.Singleton(),
 		namespaceDS.Singleton(),
+		imageCVEDS.Singleton(),
 		schema,
 	)
 }
